created a common componet for forgot password
